Refactoring ShortSeq128 to store packed sequences in a 2 element uint64_t array. The prior method of using a uint128_t just didn't confer any kind of advantage; I was basically treating it as a 2 element uint64_t array anyway, but with extra shifting and risky assumptions about system endianness.
One drawback of this refactor is that _marshall_bytes_128() is now an anti-pattern. Fixed-length non-pointer array return types don't exist in C++, so at the very least the _packed array must be constructed and passed into the function. For brevity, I chose to just pass in the entire ShortSeq128 object. I still really don't like this and I think a refactor of _new() and ShortSeq construction is in the near future.